requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
azure-core==1.30.2
azure-core-tracing-opentelemetry==1.0.0b11
azure-monitor-opentelemetry==1.6.1
azure-monitor-opentelemetry-exporter==1.0.0b28
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
fastapi==0.112.2
fixedint==0.1.6
h11==0.14.0
idna==3.8
importlib_metadata==8.0.0
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
opentelemetry-api==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-dbapi==0.47b0
opentelemetry-instrumentation-django==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-flask==0.47b0
opentelemetry-instrumentation-psycopg2==0.47b0
opentelemetry-instrumentation-requests==0.47b0
opentelemetry-instrumentation-urllib==0.47b0
opentelemetry-instrumentation-urllib3==0.47b0
opentelemetry-instrumentation-wsgi==0.47b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
packaging==24.1
psutil==5.9.8
pydantic==2.8.2
pydantic_core==2.20.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
wrapt==1.16.0
zipp==3.20.1
